refactor: create file films and trasfer films data from consts
